Fix vulnerability.
