+++ pyproject.toml
@@ -30,7 +30,6 @@ dependencies = [
     "grpcio-health-checking>=1.30.0",
     "grpcio-status>=1.26.0",
     "grpcio-reflection>=1.30.0",
-    "nltk>=3.9.3",
     "numpy>=1.14.0,<3.0.0",
     "pandas>=1.1.0,<3.0.0",
     "pandas-stubs>=1.1.0,<3.0.0",
@@ -46,6 +45,7 @@ include = ["src/ansys/fluent/core/generated/"]
 reader = ["h5py>=3.15.1"]
 ui-jupyter = ["ipywidgets>=8.1.8"]
 ui = ["panel>=1.8.1"]
+search = ["nltk>=3.9.3"]
 tests = [
     "pytest==9.0.3",
     "pytest-cov==7.1.0",
